Refactoring code through modules creation and add_if function reorganization
