+++ bmfuncts/__init__.py
@@ -11,7 +11,8 @@ from bmfuncts.employees_globals import *
 from bmfuncts.pub_globals import *
 from bmfuncts.rename_cols import *
 from bmfuncts.build_otps_info import *
-from bmfuncts.add_otps_functs import *
+from bmfuncts.add_otps import *
+from bmfuncts.add_ifs import *
 from bmfuncts.use_homonyms import *
 from bmfuncts.use_otps import *
 from bmfuncts.update_employees import *
