+++ bmgui/analyze_corpus_page.py
@@ -16,13 +16,12 @@ from tkinter import ttk
 # Local imports
 import bmfuncts.pub_globals as pg
 import bmgui.gui_globals as gg
-from bmfuncts.config_utils import set_org_params
-from bmfuncts.consolidate_pub_list import get_if_db
+from bmfuncts.add_ifs import get_if_db
 from bmfuncts.authors_analysis import authors_analysis
+from bmfuncts.config_utils import set_org_params
 from bmfuncts.coupling_analysis import coupling_analysis
 from bmfuncts.impact_factors_analysis import if_analysis
 from bmfuncts.keywords_analysis import keywords_analysis
-from bmgui.gui_globals import GUI_BUTTONS
 from bmgui.gui_utils import disable_buttons
 from bmgui.gui_utils import enable_buttons
 from bmgui.gui_utils import font_size
@@ -343,7 +342,7 @@ def create_analysis(self, master, page_name, institute, bibliometer_path, dataty
                                             variable_years,
                                             *master.years_list)
     self.OptionButton_years.config(font=self.font_OptionButton_years)
-    GUI_BUTTONS.append(self.OptionButton_years)
+    gg.GUI_BUTTONS.append(self.OptionButton_years)
 
     # - Creating year selection label
     self.font_Label_years = tkFont.Font(family=gg.FONT_NAME,
@@ -402,7 +401,7 @@ def create_analysis(self, master, page_name, institute, bibliometer_path, dataty
                                           text=launch_text,
                                           font=if_analysis_launch_font,
                                           command= _start_launch_if_analysis_try)
-    GUI_BUTTONS.append(if_analysis_launch_button)
+    gg.GUI_BUTTONS.append(if_analysis_launch_button)
     place_bellow(if_analysis_label,
                  if_analysis_launch_button,
                  dx=launch_dx_px,
@@ -462,7 +461,7 @@ def create_analysis(self, master, page_name, institute, bibliometer_path, dataty
                                           text = launch_text,
                                           font = co_analysis_launch_font,
                                           command = _start_launch_coupling_analysis_try)
-    GUI_BUTTONS.append(co_analysis_launch_button)
+    gg.GUI_BUTTONS.append(co_analysis_launch_button)
     place_bellow(co_analysis_label,
                  co_analysis_launch_button,
                  dx=launch_dx_px,
@@ -492,7 +491,7 @@ def create_analysis(self, master, page_name, institute, bibliometer_path, dataty
                                           text=launch_text,
                                           font=kw_analysis_launch_font,
                                           command= _start_launch_kw_analysis_try)
-    GUI_BUTTONS.append(kw_analysis_launch_button)
+    gg.GUI_BUTTONS.append(kw_analysis_launch_button)
     place_bellow(kw_analysis_label,
                  kw_analysis_launch_button,
                  dx=launch_dx_px,
